Prevent BlockCounter label from overflowing
Prevent BlockCounter label from spilling out of BlockCounter inner
"black" content.

This change was written on a laptop using a 1280x776 configuration (as
reported by `xrandr`).

To see this change in and out of action, it might be necessary not to
simply rebuild, but also explicitly removing src/qml/qrc_bitcoin.cpp
and/or src/qt/qrc_bitcoin.cpp.