qml: add node shutdownRequested detection
Adds a function to our NodeModel QObject that asks the underlying
m_node if shutdownRequested has been set. If it has been set then it
calls startNodeShutdown, which emits requestedShutdown. The emitted
signal can then be used to quit the GUI when a user has attempted to
close the application through a terminal.