qml: entangle the About and Developer options pages
Currently, the parent SwipeView of the AboutOptions page needs to have
an id of introductions. We abuse the knowledge that in our codebase we
will have that available, but this limits when and where the
AboutOptions can be used. This entangles the About and Developer options
so that we can avoid this, and have these pages truly be reusable.

Co-Authored-By: Johnny Carlson <johncarlson@pm.me>