qml: allow ContinueButton to stay at the bottom
With the ScrollView now containing the InformationPages, the
ContinueButton was not able to stick to the bottom of the view
due to contentHeight being too limited. This allows contentHeight
to grow with the screen while limiting it so the button doesn't
overlap with the information.

Github-Pull: bitcoin-core/gui-qml#194
Rebased-From: d17e30961ed3d0c8596ce0a8f00cb94750507e2c