qml: handle pre & header sync, move verification progress along
This connects us to the header tip signal so that we can handle when the
node is in pre & or headers sync. This calculates the progress of the
pre or headers sync and adds it to our verification progress. Pre and 
headers sync take up 1% of verification progress each, then IBD begins
at 2% progress as displayed on the block clock.