qml, desktop: Fix custom datadir doubleclick
This is a desktop-only (not mobile) workaround to disable doubleclick
on custom datadir which is on the StorageLocations.qml componennt.

Double-clicking was causing the file dialog to get open but losing
focus and moving to the background, bringing upfront the main
window which was useless because the file dialog is modal.