qml: use signal based navigation in CreateWallet pages, introduce wizard
Makes our CreateWallet flow into a wizard, as that is what it should be.
The CreateWalletWizard is introduced as a StackView with the associated
pages implementing signal based navigation.

Previously, there was an interlinked dependency between CreateName and
CreatePassword in that CreateName would pass a string for walletName
to CreatePassword. This dependency is removed by having CreateName set
the string in a property contained in the CreateWalletWizard StackView,
and CreatePassword pulling in this property to satisfy its requirement
for walletName. This is a temporary workaround so that we can still
have clearly contained pages here. A follow-up should move this into a
more appropriate backend object