qml: use signal based navigation in settings with subpages
Here, pages are clearly contained, but the logic of moving to
SettingsDeveloper from within SettingsAbout and SettingsProxy from
within SettingsConnection is still contained here within the definitions
of SettingsAbout and SettingsConnection. While all pages are
independently still usable, we do want to iron out this one kink in a
follow-up.