qml: have NodeModel compute estimating state and format sync time
Instead of having the UI construct the formatted version of the
remaining sync time as well as computing when the node is estimating the
remaining sync time using js, move this construction and computation
into the nodemodel; where it belongs.

The UI is the UI, and the backend is the backend.