qml: Introduce the CoinSelection page
The CoinSelection page is opened up from the Send page. The
button to push CoinSelection on the Send page stack is enabled
in the ellipsis menu. The toggle in the ellipsis menu can set a
QSetting for enabled/disabled. All of these components are added
to the DesktopWallets and Send page.