temporarily revert PeersNumByType (needs follow-up)
This is a reversion of
16f77a59dd4fab8890d89979464b084dc15d1466
from https://github.com/bitcoin-core/gui-qml/pull/359
which changed code in Bitcoin Core to notify new peers by their type.

That code needs to be either merged upstream to Bitcoin Core or
otherwise worked-around in gui-qml. Since the actual notification from
Bitcoin Core only sends an int of the TOTAL number of connections
(not just OUTBOUND) whatever UI there is to display this number to the
user is probably now labeled incorrectly.